refactoring: typo in package
