+++ server/src/main/kotlin/com/handcontrol/server/mqtt/MqttProperties.kt
@@ -1,4 +1,4 @@
-package com.handcontrol.server.mqqt
+package com.handcontrol.server.mqtt
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.ConstructorBinding
